demo/src/test/java/com/example/service/UserServiceTest.java
package com.example.service;

import com.example.domain.User;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;

import static org.junit.Assert.*;

/**
 * Created by hx on 2017/11/4.
 */
@RunWith(SpringRunner.class)
@SpringBootTest
public class UserServiceTest {
    @Autowired
    private UserService userService;
    @Test
    public void findOne() throws Exception {
        User user = userService.findOne(3);
        Assert.assertEquals(new Integer(20),user.getAge());

    }

}